fix(protocol): Extract out Disconnect packet logic similar to other packets
