+++ rumqttd/src/protocol/mod.rs
@@ -39,7 +39,7 @@ pub enum Packet {
     PubComp(PubComp, Option<PubCompProperties>),
     Unsubscribe(Unsubscribe),
     UnsubAck(UnsubAck),
-    Disconnect,
+    Disconnect(Disconnect, Option<DisconnectProperties>),
 }
 
 //--------------------------- Connect packet -------------------------------
@@ -381,69 +381,6 @@ pub struct UnsubAckProperties {
     pub user_properties: Vec<(String, String)>,
 }
 
-//--------------------------- Disconnect packet -------------------------------
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum DisconnectReasonCode {
-    /// Close the connection normally. Do not send the Will Message.
-    NormalDisconnection,
-    /// The Client wishes to disconnect but requires that the Server also publishes its Will Message.
-    DisconnectWithWillMessage,
-    /// The Connection is closed but the sender either does not wish to reveal the reason, or none of the other Reason Codes apply.
-    UnspecifiedError,
-    /// The received packet does not conform to this specification.
-    MalformedPacket,
-    /// An unexpected or out of order packet was received.
-    ProtocolError,
-    /// The packet received is valid but cannot be processed by this implementation.
-    ImplementationSpecificError,
-    /// The request is not authorized.
-    NotAuthorized,
-    /// The Server is busy and cannot continue processing requests from this Client.
-    ServerBusy,
-    /// The Server is shutting down.
-    ServerShuttingDown,
-    /// The Connection is closed because no packet has been received for 1.5 times the Keepalive time.
-    KeepAliveTimeout,
-    /// Another Connection using the same ClientID has connected causing this Connection to be closed.
-    SessionTakenOver,
-    /// The Topic Filter is correctly formed, but is not accepted by this Sever.
-    TopicFilterInvalid,
-    /// The Topic Name is correctly formed, but is not accepted by this Client or Server.
-    TopicNameInvalid,
-    /// The Client or Server has received more than Receive Maximum publication for which it has not sent PUBACK or PUBCOMP.
-    ReceiveMaximumExceeded,
-    /// The Client or Server has received a PUBLISH packet containing a Topic Alias which is greater than the Maximum Topic Alias it sent in the CONNECT or CONNACK packet.
-    TopicAliasInvalid,
-    /// The packet size is greater than Maximum Packet Size for this Client or Server.
-    PacketTooLarge,
-    /// The received data rate is too high.
-    MessageRateTooHigh,
-    /// An implementation or administrative imposed limit has been exceeded.
-    QuotaExceeded,
-    /// The Connection is closed due to an administrative action.
-    AdministrativeAction,
-    /// The payload format does not match the one specified by the Payload Format Indicator.
-    PayloadFormatInvalid,
-    /// The Server has does not support retained messages.
-    RetainNotSupported,
-    /// The Client specified a QoS greater than the QoS specified in a Maximum QoS in the CONNACK.
-    QoSNotSupported,
-    /// The Client should temporarily change its Server.
-    UseAnotherServer,
-    /// The Server is moved and the Client should permanently change its server location.
-    ServerMoved,
-    /// The Server does not support Shared Subscriptions.
-    SharedSubscriptionNotSupported,
-    /// This connection is closed because the connection rate is too high.
-    ConnectionRateExceeded,
-    /// The maximum connection time authorized for this connection has been exceeded.
-    MaximumConnectTime,
-    /// The Server does not support Subscription Identifiers; the subscription is not accepted.
-    SubscriptionIdentifiersNotSupported,
-    /// The Server does not support Wildcard subscription; the subscription is not accepted.
-    WildcardSubscriptionsNotSupported,
-}
-
 //--------------------------- Ping packet -------------------------------
 
 struct Ping;
@@ -531,6 +468,133 @@ pub struct PubRelProperties {
     pub user_properties: Vec<(String, String)>,
 }
 
+//------------------------------------------------------------------------
+
+//--------------------------- Disconnect packet -------------------------------
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Disconnect {
+    /// Disconnect Reason Code
+    pub reason_code: DisconnectReasonCode,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[repr(u8)]
+pub enum DisconnectReasonCode {
+    /// Close the connection normally. Do not send the Will Message.
+    NormalDisconnection = 0x00,
+    /// The Client wishes to disconnect but requires that the Server also publishes its Will Message.
+    DisconnectWithWillMessage = 0x04,
+    /// The Connection is closed but the sender either does not wish to reveal the reason, or none of the other Reason Codes apply.
+    UnspecifiedError = 0x80,
+    /// The received packet does not conform to this specification.
+    MalformedPacket = 0x81,
+    /// An unexpected or out of order packet was received.
+    ProtocolError = 0x82,
+    /// The packet received is valid but cannot be processed by this implementation.
+    ImplementationSpecificError = 0x83,
+    /// The request is not authorized.
+    NotAuthorized = 0x87,
+    /// The Server is busy and cannot continue processing requests from this Client.
+    ServerBusy = 0x89,
+    /// The Server is shutting down.
+    ServerShuttingDown = 0x8B,
+    /// The Connection is closed because no packet has been received for 1.5 times the Keepalive time.
+    KeepAliveTimeout = 0x8D,
+    /// Another Connection using the same ClientID has connected causing this Connection to be closed.
+    SessionTakenOver = 0x8E,
+    /// The Topic Filter is correctly formed, but is not accepted by this Sever.
+    TopicFilterInvalid = 0x8F,
+    /// The Topic Name is correctly formed, but is not accepted by this Client or Server.
+    TopicNameInvalid = 0x90,
+    /// The Client or Server has received more than Receive Maximum publication for which it has not sent PUBACK or PUBCOMP.
+    ReceiveMaximumExceeded = 0x93,
+    /// The Client or Server has received a PUBLISH packet containing a Topic Alias which is greater than the Maximum Topic Alias it sent in the CONNECT or CONNACK packet.
+    TopicAliasInvalid = 0x94,
+    /// The packet size is greater than Maximum Packet Size for this Client or Server.
+    PacketTooLarge = 0x95,
+    /// The received data rate is too high.
+    MessageRateTooHigh = 0x96,
+    /// An implementation or administrative imposed limit has been exceeded.
+    QuotaExceeded = 0x97,
+    /// The Connection is closed due to an administrative action.
+    AdministrativeAction = 0x98,
+    /// The payload format does not match the one specified by the Payload Format Indicator.
+    PayloadFormatInvalid = 0x99,
+    /// The Server has does not support retained messages.
+    RetainNotSupported = 0x9A,
+    /// The Client specified a QoS greater than the QoS specified in a Maximum QoS in the CONNACK.
+    QoSNotSupported = 0x9B,
+    /// The Client should temporarily change its Server.
+    UseAnotherServer = 0x9C,
+    /// The Server is moved and the Client should permanently change its server location.
+    ServerMoved = 0x9D,
+    /// The Server does not support Shared Subscriptions.
+    SharedSubscriptionNotSupported = 0x9E,
+    /// This connection is closed because the connection rate is too high.
+    ConnectionRateExceeded = 0x9F,
+    /// The maximum connection time authorized for this connection has been exceeded.
+    MaximumConnectTime = 0xA0,
+    /// The Server does not support Subscription Identifiers; the subscription is not accepted.
+    SubscriptionIdentifiersNotSupported = 0xA1,
+    /// The Server does not support Wildcard subscription; the subscription is not accepted.
+    WildcardSubscriptionsNotSupported = 0xA2,
+}
+
+impl TryFrom<u8> for DisconnectReasonCode {
+    type Error = Error;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        let rc = match value {
+            0x00 => Self::NormalDisconnection,
+            0x04 => Self::DisconnectWithWillMessage,
+            0x80 => Self::UnspecifiedError,
+            0x81 => Self::MalformedPacket,
+            0x82 => Self::ProtocolError,
+            0x83 => Self::ImplementationSpecificError,
+            0x87 => Self::NotAuthorized,
+            0x89 => Self::ServerBusy,
+            0x8B => Self::ServerShuttingDown,
+            0x8D => Self::KeepAliveTimeout,
+            0x8E => Self::SessionTakenOver,
+            0x8F => Self::TopicFilterInvalid,
+            0x90 => Self::TopicNameInvalid,
+            0x93 => Self::ReceiveMaximumExceeded,
+            0x94 => Self::TopicAliasInvalid,
+            0x95 => Self::PacketTooLarge,
+            0x96 => Self::MessageRateTooHigh,
+            0x97 => Self::QuotaExceeded,
+            0x98 => Self::AdministrativeAction,
+            0x99 => Self::PayloadFormatInvalid,
+            0x9A => Self::RetainNotSupported,
+            0x9B => Self::QoSNotSupported,
+            0x9C => Self::UseAnotherServer,
+            0x9D => Self::ServerMoved,
+            0x9E => Self::SharedSubscriptionNotSupported,
+            0x9F => Self::ConnectionRateExceeded,
+            0xA0 => Self::MaximumConnectTime,
+            0xA1 => Self::SubscriptionIdentifiersNotSupported,
+            0xA2 => Self::WildcardSubscriptionsNotSupported,
+            other => return Err(Error::InvalidConnectReturnCode(other)),
+        };
+
+        Ok(rc)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DisconnectProperties {
+    /// Session Expiry Interval in seconds
+    pub session_expiry_interval: Option<u32>,
+
+    /// Human readable reason for the disconnect
+    pub reason_string: Option<String>,
+
+    /// List of user properties
+    pub user_properties: Vec<(String, String)>,
+
+    /// String which can be used by the Client to identify another Server to use.
+    pub server_reference: Option<String>,
+}
 //------------------------------------------------------------------------
 
 /// Quality of service
