+++ rumqttd/src/protocol/v4/mod.rs
@@ -299,7 +299,12 @@ impl Protocol for V4 {
             return match packet_type {
                 PacketType::PingReq => Ok(Packet::PingReq(PingReq)),
                 PacketType::PingResp => Ok(Packet::PingResp(PingResp)),
-                PacketType::Disconnect => Ok(Packet::Disconnect),
+                PacketType::Disconnect => Ok(Packet::Disconnect(
+                    Disconnect {
+                        reason_code: DisconnectReasonCode::NormalDisconnection,
+                    },
+                    None,
+                )),
                 _ => Err(Error::PayloadRequired),
             };
         }
@@ -326,7 +331,9 @@ impl Protocol for V4 {
             PacketType::PubRec => Packet::PubRec(pubrec::read(fixed_header, packet)?, None),
             PacketType::PubRel => Packet::PubRel(pubrel::read(fixed_header, packet)?, None),
             PacketType::PubComp => Packet::PubComp(pubcomp::read(fixed_header, packet)?, None),
-            PacketType::Disconnect => Packet::Disconnect,
+            // v4 Disconnect packet gets handled in the previous check, this branch gets hit when
+            // Disconnect packet has properties which is only valid for v5
+            PacketType::Disconnect => return Err(Error::InvalidProtocol),
             _ => unreachable!(),
         };
