+++ rumqttd/src/protocol/v5/mod.rs
@@ -371,7 +371,12 @@ impl Protocol for V5 {
             return match packet_type {
                 PacketType::PingReq => Ok(Packet::PingReq(PingReq)),
                 PacketType::PingResp => Ok(Packet::PingResp(PingResp)),
-                PacketType::Disconnect => Ok(Packet::Disconnect),
+                PacketType::Disconnect => Ok(Packet::Disconnect(
+                    Disconnect {
+                        reason_code: DisconnectReasonCode::NormalDisconnection,
+                    },
+                    None,
+                )),
                 _ => Err(Error::PayloadRequired),
             };
         }
@@ -401,7 +406,10 @@ impl Protocol for V5 {
             }
             PacketType::PingReq => Packet::PingReq(PingReq),
             PacketType::PingResp => Packet::PingResp(PingResp),
-            PacketType::Disconnect => Packet::Disconnect,
+            PacketType::Disconnect => {
+                let (disconnect, properties) = disconnect::read(fixed_header, packet)?;
+                Packet::Disconnect(disconnect, properties)
+            }
             _ => unreachable!(),
         };
